Do not humanizePluginName the user-provided name of an app
